requirements.txt
Flask>=2.0.0
google-generativeai>=0.8.0
transformers>=4.55.0
torch>=2.0.0
torchaudio>=2.0.0
scipy>=1.13.0
numpy>=1.26.0